Update FeatureOfInterestConstraint.ttl
Rui-- i corrected a small spelling error, but also have a general question.  My knowledge of SHACL and SOSA are not as good as yours, but I imagined that SHACL could be used to validate that any SOSA:Observation had at least a specification of what is the SOSA:FeatureOfInterest as well as its SOSA:ObservedProperty.  This is where it would help me to start with an example involving actual sample Datasets, beginning with their native format -- reviewing values (and metadata) from the native format are materialized as triples, and then examining how these were "aligned with SOSA" -- might clarify the utility of some SHACL constraint on those triples that we  "aligned with SOSA".  Do we have a small sample of data (raw, triplified, and SOSA-aligned) that i could examine to get a better feeling for how some SHACL validation might be especially useful?  For example,  I am not sure that requiring 'SOSA:hasSample' is as essential as validating that some SOSA:FeatureOfInterest has been specified for every SOSA:Observation.  I apologize if I am simply showing my own ignorance here!!!